Added price category entity

+++ src/main/java/com/alex/rest/resources/PriceResource.java
@@ -2,14 +2,17 @@ package com.alex.rest.resources;
 
 import com.alex.rest.domen.Price;
 import com.alex.rest.domen.Product;
+import com.alex.rest.repository.payment.PriceRepository;
 import com.alex.rest.service.PriceService;
 import com.alex.rest.service.ProductService;
 import com.alex.rest.service.QueryService;
+import com.alex.rest.service.exceptions.InvalidParameterException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 import java.util.Collection;
 import java.util.Optional;
 
@@ -18,35 +21,31 @@ import java.util.Optional;
 @Component
 public class PriceResource {
 
-    private PriceService priceService;
-
-    private ProductService productService;
-
     @Autowired
-    public PriceResource(PriceService priceService, ProductService productService) {
-        this.priceService = priceService;
-        this.productService = productService;
-    }
+    private PriceService priceService;
+    @Autowired
+    private ProductService productService;
+    @Autowired
+    private PriceRepository priceRepository;
 
     // Retrieve
-    @GET
-    @Path("/*")
-    public Collection<Price> findAll() {
-        return priceService.findAll();
-    }
 
     @GET
-    public Price getPrice(@PathParam("productId") Long id) {
-//        Optional<Price> matchingPrice = priceService.findAll().stream()
-//                .filter(p -> p.getProduct().getPrice().getId().equals(id))
-//                .findFirst();
-        return priceService.findById(1L);
+    public Price getPrice(@PathParam("productId") Long id) throws InvalidParameterException {
+        if (id < 0) {
+            throw new InvalidParameterException("Invalid input");
+        }
+        Optional<Price> price = Optional.ofNullable(productService.findById(id).getPrice());
+        return price.orElseThrow(() -> new NullPointerException("The price is not found."));
     }
 
     // Create
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
-    public void createPrice(@PathParam("productId") Long id, Price price) {
+    public void createPrice(@PathParam("productId") Long id, Price price) throws InvalidParameterException {
+        if (id < 0) {
+            throw new InvalidParameterException("Invalid input");
+        }
         Product searchedProduct = productService.findById(id);
         price.setProduct(searchedProduct);
         priceService.add(price);
@@ -54,7 +53,11 @@ public class PriceResource {
 
     @DELETE
     @Path("/{id}")
-    public void delete(@PathParam("id") long id) {
+    public Response delete(@PathParam("id") long id) {
+        if (!priceRepository.isExist(id)) {
+            throw new NullPointerException("The price with id = " + id + "is not exist!");
+        }
         priceService.delete(id);
+        return Response.ok("The price is deleted successfully").type(MediaType.TEXT_PLAIN_TYPE).build();
     }
 }
